% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_directory.R
\name{read_directory}
\alias{read_directory}
\title{Read delimited files and/or rds files from a directory}
\usage{
read_directory(directory_path)
}
\arguments{
\item{directory_path}{Absolute path of directory}
}
\description{
Read delimited files and/or rds files from a directory
}
