% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_project.R
\name{setup_project}
\alias{setup_project}
\title{Setup data science project in R project folder}
\usage{
setup_project(path = ".", project_type = "ds", simple = TRUE)
}
\arguments{
\item{path}{The location to create project skeleton. If using inside R project, will create in parent directory of project.}

\item{project_type}{The type of project skeleton to create.}

\item{simple}{If TRUE, the default, builds simpler skeleton with less directories.}
}
\description{
Setup data science project in R project folder
}
\examples{
\dontrun{read_directory()}
}
